Simplify ReloadFromDiskUtils
Turns out the the implementation of a reload from disk functionality can
be largely simplified and optimized. The method Model.loadRad(...) that is
the standard for opening a Mastodon Model, can also be used for reloading
the Model. This should also be way more robust.
