+++ src/test/java/org/mastodon/mamut/collaboration/utils/ReloadFromDiskUtilsTest.java
@@ -45,21 +45,4 @@ public class ReloadFromDiskUtilsTest
 			}
 		}
 	}
-
-	@Test
-	public void testOutgoingEdgesOrder()
-	{
-		ModelGraph graphA = new ModelGraph();
-		Spot a = graphA.addVertex().init( 0, new double[ 3 ], 1 );
-		a.setLabel( "A" );
-		Spot b = graphA.addVertex().init( 1, new double[ 3 ], 1 );
-		b.setLabel( "B" );
-		Spot c = graphA.addVertex().init( 1, new double[ 3 ], 1 );
-		c.setLabel( "C" );
-		graphA.addEdge( a, c );
-		graphA.addEdge( a, b );
-		ModelGraph graphB = new ModelGraph();
-		ReloadFromDiskUtils.copyGraphFromTo( graphA, graphB );
-		ModelAsserts.assertGraphEquals( graphB, graphA );
-	}
 }
